ADDED content page column
The last url in the transactionPath is the content page by definition. These are all extracted into a separate column and the transactionPath is removed. This is believed to increase clustering accuracy due to the single url comparison, and the context from the transaction is still kept in the keywords column in condensed form.

Also REFACTORED JsonProcessor, added the extra step and cleaned up some of the new additions.